requirements.txt
langchain-openai
langchain_core
python-dotenv
streamlit
langchain-ollama
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
langchain_groq
sentence_transformers